+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 * Add export excel feature for list components.
+* Add `fixed` settings for columns in `developerUserSettings` for list components.
+
+### Changed
+* Update `jquery.colResizable` plugin to version 1.6.
 
 ## [0.9.0-beta.0] - 2017-02-22
 ### Added